resources/setuptools/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "py-project"
version = "0.0.1"
dependencies = [
    "PyYAML==6.0.1"
]

[tool.setuptools.packages.find]
# All the following settings are optional:
where = ["src"]  # ["."] by default
include = ["py_project*"]  # ["*"] by default
exclude = ["py_project.tests*"]  # empty by default

[project.scripts]
py-project = "py_project.main:main"
